Be consistent with property names. Dashes, not underscores. 	modified:   src/y-axis-view.c 	modified:   src/y-density-plot.c 	modified:   src/y-density-plot.h 	modified:   src/y-plot-widget.c 	modified:   src/y-scatter-view.c

+++ src/y-density-plot.h
@@ -53,7 +53,7 @@ struct _YDensityPlot {
   
   gboolean draw_line;
   GtkOrientation line_dir;
-  int line_pos, line_width;
+  double line_pos, line_width;
   
   gboolean draw_dot;
   double dot_pos_x, dot_pos_y;
